pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "task_manager"
authors = ["krushovice <kickstar69@yandex.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.8.1"
python-dotenv = "0.21.0"
pytest-cov = "^4.1.0"
beautifulsoup4 = "4.12.2"
postgres = "^4.0"
flake8 = "^6.1.0"
psycopg2-binary = "^2.9.9"
django = "^4.2.7"
dj-database-url = "^2.1.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^21.2.0"
django-bootstrap5 = "^23.3"
django-filter = "^23.5"
rollbar = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
